tests/integration/test_s3_lambda.py
"""Integration tests for S3 to Lambda trigger - Phase 2+."""

import pytest


@pytest.mark.skip(reason="Phase 2 - S3 Lambda integration not yet implemented")
def test_s3_upload_triggers_lambda() -> None:
    """Test that uploading to S3 triggers the Lambda function."""
    pass


@pytest.mark.skip(reason="Phase 2 - S3 Lambda integration not yet implemented")
def test_lambda_receives_correct_event() -> None:
    """Test that Lambda receives correct S3 event structure."""
    pass
