requirements.txt
aws-cdk-lib>=2.232.2,<3.0.0
constructs>=10.0.0,<11.0.0

# Phase 4: Database + Web Portal
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
streamlit>=1.30.0
boto3>=1.28.0
python-dotenv>=1.0.0

# Phase 6: Analytics Dashboards
plotly>=5.18.0
